Split file to Util.swift
